test(ui): Convert integration row to RTL
